Started working on events page
