+++ package.json
@@ -14,7 +14,8 @@
     "react-dom": "^18.2.0",
     "react-fast-marquee": "^1.6.0",
     "react-helmet-async": "^1.3.0",
-    "react-icons": "^4.11.0"
+    "react-icons": "^4.11.0",
+    "react-router-dom": "^7.6.2"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
