+++ src/Pages/Components/Events/Events.module.css
@@ -1,0 +1,24 @@
+.buttonContainer {
+    display: flex;
+    justify-content: center;
+    gap: 1rem;
+    overflow-x: auto;
+    padding-bottom: 1rem;
+    scrollbar-width: none;
+    -ms-overflow-style: none;
+}
+
+.buttonContainer::-webkit-scrollbar {
+    display: none;
+}
+
+.eventButton {
+    padding: 0.5rem 1.5rem;
+    border-radius: 9999px;
+    min-width: 100px;
+    transition: all 0.3s ease;
+    cursor: pointer;
+    border: none;
+    background-color: #ae59ff;
+    color: white;
+}
