+++ src/Pages/Components/ExploreLC/ExploreLC.tsx
