+++ src/Pages/Components/Footer/Footer.tsx
@@ -1,7 +1,7 @@
 import styles from "./Footer.module.css";
 import { Instagram, X, Youtube, LinkedIn } from "./assets/svg";
-import { ULearnWhite } from "../../assets/svg/svg";
-import data from "../../../data.json";
+import { ULearnWhite } from "../../../assets/svg/svg";
+import data from "../../../../data.json";
 
 const Footer = () => {
   return (
