+++ src/Pages/Components/Hero/assets/svg.tsx
