+++ src/Pages/Components/Team/Team.tsx
@@ -1,5 +1,5 @@
 import styles from "./Team.module.css";
-import data from "../../../data.json";
+import data from "../../../../data.json";
 
 const Team = () => {
   return (
