+++ src/Pages/Events.module.css
@@ -1,0 +1,419 @@
+.events {
+    width: 100%;
+    padding: 6rem 2rem;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    gap: 2rem;
+}
+
+.events h2 {
+    text-align: center;
+    align-self: center;
+    color: #ae59ff;
+    font-size: 2rem;
+    margin-bottom: 1rem;
+}
+
+/* Hero Section */
+.heroSection {
+    width: 100%;
+    max-width: 1200px;
+    margin: 0 auto;
+    position: relative;
+}
+
+.heroCard {
+    background: mediumpurple;
+    border-radius: 2rem;
+    padding: 3rem;
+    color: white;
+    position: relative;
+    overflow: hidden;
+    min-height: 400px;
+    transition: opacity 0.3s ease-in-out;
+}
+
+.heroCard::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: rgba(0, 0, 0, 0.2);
+    border-radius: 2rem;
+}
+
+.heroContent {
+    display: flex;
+    align-items: center;
+    gap: 2rem;
+    position: relative;
+    z-index: 2;
+    height: 100%;
+}
+
+.heroLeft {
+    flex: 1;
+    display: flex;
+    gap: 2rem;
+    align-items: center;
+}
+
+.heroDate {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background: rgba(255, 255, 255, 0.2);
+    border-radius: 1rem;
+    padding: 1.5rem 1rem;
+    min-width: 100px;
+    backdrop-filter: blur(10px);
+}
+
+.heroMonth {
+    font-size: 1.2rem;
+    font-weight: bold;
+    text-transform: uppercase;
+    letter-spacing: 1px;
+}
+
+.heroDay {
+    font-size: 3rem;
+    font-weight: bold;
+    line-height: 1;
+}
+
+.heroText {
+    flex: 1;
+}
+
+.heroTitle {
+    font-size: 2.5rem;
+    font-weight: bold;
+    margin-bottom: 1rem;
+    line-height: 1.2;
+}
+
+.heroDescription {
+    font-size: 1.2rem;
+    line-height: 1.6;
+    opacity: 0.9;
+}
+
+.heroImageContainer {
+    flex: 0 0 300px;
+}
+
+.heroImage {
+    width: 100%;
+    height: 250px;
+    object-fit: cover;
+    border-radius: 1rem;
+    box-shadow: 0 20px 40px rgba(0, 0, 0, 0.3);
+}
+
+/* Hero Navigation Dots */
+.heroDots {
+    display: flex;
+    justify-content: center;
+    gap: 0.5rem;
+    margin-top: 2rem;
+    position: relative;
+    z-index: 2;
+}
+
+.heroDot {
+    width: 12px;
+    height: 12px;
+    border-radius: 50%;
+    border: none;
+    background: rgba(255, 255, 255, 0.4);
+    cursor: pointer;
+    transition: all 0.3s ease;
+}
+
+.heroDot:hover {
+    background: rgba(255, 255, 255, 0.6);
+    transform: scale(1.1);
+}
+
+.heroDotActive {
+    background: white;
+    transform: scale(1.2);
+}
+
+/* Fade Animations */
+.fadeIn {
+    opacity: 1;
+}
+
+.fadeOut {
+    opacity: 0;
+}
+
+/* Year selection styles */
+.yearScroll {
+    position: relative;
+    width: 100%;
+    padding: 0 1rem;
+}
+
+.yearButtons {
+    display: flex;
+    justify-content: center;
+    gap: 1rem;
+    overflow-x: auto;
+    padding-bottom: 1rem;
+    scrollbar-width: none;
+    -ms-overflow-style: none;
+}
+
+.yearButtons::-webkit-scrollbar {
+    display: none;
+}
+
+.yearButton {
+    padding: 0.75rem 2rem;
+    border-radius: 9999px;
+    min-width: 120px;
+    transition: all 0.3s ease;
+    cursor: pointer;
+    border: none;
+    font-weight: 600;
+    font-size: 1rem;
+}
+
+.yearButton.active {
+    background-color: #ae59ff;
+    color: white;
+    transform: translateY(-2px);
+    box-shadow: 0 8px 25px rgba(174, 89, 255, 0.3);
+}
+
+.yearButton:not(.active) {
+    background-color: #f3f4f6;
+    color: #4b5563;
+}
+
+.yearButton:not(.active):hover {
+    background-color: #e5e7eb;
+    transform: translateY(-1px);
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+/* Events Grid */
+.eventsGrid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+    gap: 2rem;
+    margin-top: 1rem;
+}
+
+.eventCard {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    align-items: center;
+    background: white;
+    border-radius: 1.5rem;
+    box-shadow: 0 8px 32px rgba(0, 0, 0, 0.1);
+    padding: 1.5rem;
+    transition: all 0.3s ease;
+    cursor: pointer;
+}
+
+.eventCard:hover {
+    transform: translateY(-8px);
+    box-shadow: 0 16px 48px rgba(0, 0, 0, 0.15);
+}
+
+.eventContent {
+    display: flex;
+    width: 100%;
+    justify-content: space-between;
+    gap: 1.5rem;
+}
+
+.eventDate {
+    display: flex;
+    align-items: center;
+    color: black;
+    gap: 8px;
+    border-radius: 1rem;
+    padding: 1rem 0.75rem;
+    min-width: 80px;
+}
+
+.eventMonth {
+    font-size: 1.5rem;
+    font-weight: bold;
+    text-transform: uppercase;
+    letter-spacing: 0.5px;
+}
+
+.eventDay {
+    font-size: 1.5rem;
+    font-weight: bold;
+    line-height: 1;
+}
+
+.eventText {
+    flex: 1;
+}
+
+.eventTitle {
+    font-size: 1.3rem;
+    font-weight: bold;
+    color: #333;
+    margin-bottom: 0.5rem;
+    line-height: 1.3;
+}
+
+.eventDescription {
+    font-size: 1rem;
+    color: #666;
+    line-height: 1.5;
+    display: -webkit-box;
+    -webkit-line-clamp: 2;
+    -webkit-box-orient: vertical;
+    overflow: hidden;
+}
+
+.eventImage {
+    height: 250px;
+    width: auto;
+    object-fit: contain;
+    border-radius: 1rem;
+    margin-left: 1rem;
+    flex-shrink: 0;
+}
+
+/* Responsive Design */
+@media (max-width: 1200px) {
+    .heroContent {
+        flex-direction: column;
+        text-align: center;
+    }
+
+    .heroLeft {
+        flex-direction: column;
+        align-items: center;
+    }
+
+    .heroImageContainer {
+        flex: none;
+        width: 100%;
+        max-width: 400px;
+    }
+
+    .heroImage {
+        height: 200px;
+    }
+}
+
+@media (max-width: 880px) {
+    .events {
+        padding: 0 1rem;
+    }
+
+    .heroCard {
+        padding: 2rem;
+        min-height: 300px;
+    }
+
+    .heroTitle {
+        font-size: 2rem;
+    }
+
+    .heroDescription {
+        font-size: 1rem;
+    }
+
+    .eventsGrid {
+        grid-template-columns: 1fr;
+        gap: 1.5rem;
+    }
+
+    .eventCard {
+        padding: 1rem;
+    }
+
+    .yearButton {
+        min-width: 100px;
+        padding: 0.5rem 1.5rem;
+    }
+}
+
+@media (max-width: 500px) {
+    .heroCard {
+        padding: 1.5rem;
+    }
+
+    .heroTitle {
+        font-size: 1.5rem;
+    }
+
+    .heroDescription {
+        font-size: 0.9rem;
+    }
+
+    .heroDate {
+        padding: 1rem 0.75rem;
+        min-width: 80px;
+    }
+
+    .heroDay {
+        font-size: 2rem;
+    }
+
+    .eventCard {
+        flex-direction: column;
+        text-align: center;
+        gap: 1rem;
+    }
+
+    /* .eventContent {
+        flex-direction: column;
+        gap: 1rem;
+    } */
+
+    .eventMonth {
+        font-size: 1rem;
+    }
+
+    .eventDay {
+        font-size: 1rem;
+    }
+
+    .eventImage {
+        margin-left: 0;
+        width: 100%;
+        height: 150px;
+    }
+
+    .yearButton {
+        min-width: 80px;
+        padding: 0.5rem 1rem;
+        font-size: 0.9rem;
+    }
+}
+
+@media (width < 345px) {
+    .heroTitle {
+        font-size: 1.2rem;
+    }
+
+    .heroDescription {
+        font-size: 0.8rem;
+    }
+
+    .eventTitle {
+        font-size: 1.1rem;
+    }
+
+    .eventDescription {
+        font-size: 0.9rem;
+    }
+}
